Using float for line width
git-svn-id: http://mapnik-utils.googlecode.com/svn@84 dbea3d40-3153-0410-b551-57805770c6e6

+++ trunk/serverside/styles/style.py
@@ -65,7 +65,7 @@ properties = {
     'line-color': color,
 
     # 0.0 - n (default 1.0)
-    'line-width': int,
+    'line-width': float,
 
     # 0.0 - 1.0 (default 1.0)
     'line-opacity': float,
